chat-admin/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pyjwt>=2.8.0
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
greenlet>=3.0.0
python-multipart>=0.0.6
minio>=7.2.0
jinja2>=3.1.0
pydantic-settings>=2.1.0
bcrypt>=4.0.0
email-validator>=2.0.0
